Delete Owner

+++ Veterinary.Web/Controllers/PetTypesController.cs
@@ -1,5 +1,6 @@
 ﻿using Microsoft.AspNetCore.Mvc;
 using Microsoft.EntityFrameworkCore;
+using System;
 using System.Linq;
 using System.Threading.Tasks;
 using Veterinary.Web.Data.Entities;
@@ -113,7 +114,6 @@ namespace Veterinary.Web.Controllers
             return View(petType);
         }
 
-        // GET: PetTypes/Delete/5
         public async Task<IActionResult> Delete(int? id)
         {
             if (id == null)
@@ -122,24 +122,32 @@ namespace Veterinary.Web.Controllers
             }
 
             var petType = await _context.PetTypes
-                .FirstOrDefaultAsync(m => m.Id == id);
+                .Include(pt => pt.Pets)
+                .FirstOrDefaultAsync(pt => pt.Id == id);
             if (petType == null)
             {
                 return NotFound();
             }
 
-            return View(petType);
-        }
+            if (petType.Pets.Count > 0)
+            {
 
-        // POST: PetTypes/Delete/5
-        [HttpPost, ActionName("Delete")]
-        [ValidateAntiForgeryToken]
-        public async Task<IActionResult> DeleteConfirmed(int id)
-        {
-            var petType = await _context.PetTypes.FindAsync(id);
-            _context.PetTypes.Remove(petType);
-            await _context.SaveChangesAsync();
-            return RedirectToAction(nameof(Index));
+                ModelState.AddModelError(string.Empty, "The pet type can´t be removed ");
+                return RedirectToAction(nameof(Index));
+            }
+
+            try
+            {
+                _context.PetTypes.Remove(petType);
+                await _context.SaveChangesAsync();
+                return RedirectToAction(nameof(Index));
+            }
+            catch (Exception ex)
+            {
+
+                ModelState.AddModelError(string.Empty, ex.ToString());
+                return RedirectToAction(nameof(Index));
+            }
         }
 
         private bool PetTypeExists(int id)
